third_party/blink/web_tests/accessibility/heading-level-expected.txt

This tests that headings have a level matching the implicit default value or explicitly defined aria-level value.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS: level is 1.
PASS: level is 2.
PASS: level is 3.
PASS: level is 4.
PASS: level is 5.
PASS: level is 6.
PASS: level is 1.
PASS: level is 2.
PASS: level is 3.
PASS: level is 4.
PASS: level is 5.
PASS: level is 6.
PASS: level is 1.
PASS: level is 2.
PASS: level is 3.
PASS: level is 4.
PASS: level is 5.
PASS: level is 6.
PASS: level is 1.
PASS: level is 1.
PASS: level is 2.

PASS successfullyParsed is true

TEST COMPLETE
